requirements.txt
tensorflow~=2.2.0
matplotlib~=3.1.1
numpy~=1.17.2
jupyterlab
pillow~=6.2.0